refactor(mcp): separate proxy response handling
